+++ sdmx-dl-api/src/main/java/sdmxdl/AttributeRelationship.java
@@ -1,0 +1,6 @@
+package sdmxdl;
+
+public enum AttributeRelationship {
+
+    UNKNOWN, DATAFLOW, GROUP, SERIES, OBSERVATION
+}