+++ sdmx-dl-api/src/test/java/tests/sdmxdl/api/RepoSamples.java
@@ -50,8 +50,8 @@ public class RepoSamples {
     public static final Dimension DIM2 = Dimension.builder().id("REGION").codelist(CL2).label("Region").position(3).build();
     public static final Dimension DIM3 = Dimension.builder().id("SECTOR").codelist(CL3).label("Sector").position(4).build();
 
-    public static final Attribute NOT_CODED_ATTRIBUTE = Attribute.builder().id("TITLE").label("Title").build();
-    public static final Attribute CODED_ATTRIBUTE = Attribute.builder().id("OBS_STATUS").codelist(CL4).label("Observation status").build();
+    public static final Attribute NOT_CODED_ATTRIBUTE = Attribute.builder().id("TITLE").label("Title").relationship(AttributeRelationship.SERIES).build();
+    public static final Attribute CODED_ATTRIBUTE = Attribute.builder().id("OBS_STATUS").codelist(CL4).label("Observation status").relationship(AttributeRelationship.OBSERVATION).build();
 
     public static final DataStructure STRUCT = DataStructure
             .builder()