+++ sdmx-dl-cli/src/test/java/sdmxdl/cli/ListConceptsCommandTest.java
@@ -43,9 +43,9 @@ public class ListConceptsCommandTest {
                 .isEmpty();
 
         assertThat(FileSample.readAll(out))
-                .contains("Concept,Label,Type,Coded,Index", atIndex(0))
-                .contains("FREQ,Frequency,dimension,true,0", atIndex(1))
-                .contains("TITLE_COMPL,Title complement,attribute,false,", atIndex(18))
+                .contains("Concept,Label,Type,Coded,Index,Relationship", atIndex(0))
+                .contains("FREQ,Frequency,dimension,true,0,", atIndex(1))
+                .contains("TITLE_COMPL,Title complement,attribute,false,,GROUP", atIndex(18))
                 .hasSize(19);
     }
 }