+++ sdmx-dl-format-xml/src/test/java/internal/sdmxdl/format/xml/XMLStreamStructure21Test.java
@@ -56,12 +56,12 @@ public class XMLStreamStructure21Test {
             assertThat(attributes).hasSize(11);
             assertThat(attributes).filteredOn(attr -> attr.getId().equals("TIME_FORMAT")).singleElement().satisfies(x -> {
                 assertThat(x.getLabel()).isEqualTo("Time format code");
-//                assertThat(x.getRelationShip()).isEqualTo(Attribute.RelationShip.SERIES);
+                assertThat(x.getRelationship()).isEqualTo(AttributeRelationship.SERIES);
                 assertThat(x.getCodes()).isEmpty();
             });
             assertThat(attributes).filteredOn(attr -> attr.getId().equals("OBS_STATUS")).singleElement().satisfies(x -> {
                 assertThat(x.getLabel()).isEqualTo("Observation status");
-//                assertThat(x.getRelationShip()).isEqualTo(Attribute.RelationShip.OBS);
+                assertThat(x.getRelationship()).isEqualTo(AttributeRelationship.OBSERVATION);
                 assertThat(x.getCodes()).isNotEmpty();
             });
         });